requirements.txt
django
django-environ
django-cors-headers
djangorestframework
django-admin-sortable
